pyproject.toml
[tool.poetry]
name = "tft-cli"
version = "0.0.17"
description = "Testing Farm CLI tool"
authors = ["Miroslav Vadkerti <mvadkert@redhat.com>"]
license = "Apache-2.0"

packages = [
  { include = "tft", from = "src" }
]

[tool.poetry.scripts]
testing-farm = "tft.cli.tool:app"

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.7.0"}
# click 8.1.0 broke typer 0.4.0
# https://github.com/tiangolo/typer/pull/375
click = "~8.0.4"
dynaconf = "^3.1.7"
colorama = "^0.4.4"
requests = "^2.27.1"
ruamel-yaml = "^0.18.6"

[tool.poetry.dev-dependencies]
pyre-check = "^0.9.10"
pytest = "^7.0.0"
isort = "^5.10.1"
black = "^22.1.0"
pre-commit = "^2.17.0"
tox = "^3.24.5"

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.isort]
profile = "black"
multi_line_output = 3

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
